% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsterdown-package.R
\docType{package}
\name{amsterdown-package}
\alias{amsterdown}
\alias{amsterdown-package}
\title{amsterdown: An RMarkdown PhD thesis template for the University of
    Amsterdam}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An RMarkdown PhD thesis template for the University
    of Amsterdam.  Enables reproducible thesis writing using (R)Markdown,
    with data, code, text, figures, and the like all in one place. Uses
    the bookdown package to convert the (R)Markdown files to pdf or html
    outputs. A modification of the default LaTeX template ensures the PhD
    thesis conforms to the University of Amsterdam requirements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lcreteig/amsterdown}
  \item Report bugs at \url{https://github.com/lcreteig/amsterdown/issues}
}

}
\author{
\strong{Maintainer}: Leon Reteig \email{leonreteig@gmail.com} (\href{https://orcid.org/0000-0002-5814-0992}{ORCID})

}
\keyword{internal}
